Fix incorrect and missing types: - AccountStatus.logonCount from int to long - AccountStatus.logonMinutes from int to long - APICall.groupID from int to long - CalendarEventAttendee.eventID from int to long - CharacterSheet.cloneSkillPoints from int to long - CharacterSheet.freeSkillPoints from int to long - SkillInfo.freeSkillPoints from int to long - UpcomingCalendarEvent.eventID from int to long - ContainerLog.itemTypeID from long to int - ContainerLog.locationID from int to long - ContainerLog.quantity from int to long - ContainerLog.typeID from long to int - MemberTracking.baseID from int to long - MemberTracking.locationID from int to long - Outpost.solarSystemID from long to int - Starbase.locationID from int to long - Asset.quantity from int to long - Contract.startStationID from int to long - Contract.endStationID from int to long - Contract.volume from long to double - ContractItem.rawQuantity from int to long - IndustryJob.activityID from long to int - IndustryJob.blueprintTypeID from long to int - IndustryJob.productTypeID from long to int - IndustryJob.timeInSeconds from long to int - WalletTransaction.typeID from long to int - WalletTransacton.clientTypeID missing (it’s an int)

+++ src/main/java/enterprises/orbital/evexmlapi/act/IAccountStatus.java
@@ -8,15 +8,16 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 /**
  * A response to a request for EVE Account Status.
  */
-@JsonSerialize(as = IAccountStatus.class)
+@JsonSerialize(
+    as = IAccountStatus.class)
 public interface IAccountStatus {
   public Date getPaidUntil();
 
   public Date getCreateDate();
 
-  public int getLogonCount();
+  public long getLogonCount();
 
-  public int getLogonMinutes();
+  public long getLogonMinutes();
 
   public List<IMultiCharacterTraining> getMultiCharacterTraining();
