+++ src/main/java/enterprises/orbital/evexmlapi/chr/ICalendarEventAttendee.java
@@ -2,9 +2,10 @@ package enterprises.orbital.evexmlapi.chr;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
-@JsonSerialize(as = ICalendarEventAttendee.class)
+@JsonSerialize(
+    as = ICalendarEventAttendee.class)
 public interface ICalendarEventAttendee {
-  public int getEventID();
+  public long getEventID();
 
   public long getCharacterID();
