+++ src/main/java/enterprises/orbital/impl/evexmlapi/act/AccountStatusResponse.java
@@ -11,11 +11,12 @@ import enterprises.orbital.impl.evexmlapi.ApiResponse;
 public class AccountStatusResponse extends ApiResponse implements IAccountStatus {
   private Date                                  paidUntil;
   private Date                                  createDate;
-  private int                                   logonCount;
-  private int                                   logonMinutes;
+  private long                                  logonCount;
+  private long                                  logonMinutes;
   private final List<ApiMultiCharacterTraining> multiCharacterTraining = new ArrayList<ApiMultiCharacterTraining>();
 
-  public void addMultiCharacterTraining(ApiMultiCharacterTraining training) {
+  public void addMultiCharacterTraining(
+                                        ApiMultiCharacterTraining training) {
     multiCharacterTraining.add(training);
   }
 
@@ -31,7 +32,8 @@ public class AccountStatusResponse extends ApiResponse implements IAccountStatus
     return paidUntil;
   }
 
-  public void setPaidUntil(Date paidUntil) {
+  public void setPaidUntil(
+                           Date paidUntil) {
     this.paidUntil = paidUntil;
   }
 
@@ -40,25 +42,28 @@ public class AccountStatusResponse extends ApiResponse implements IAccountStatus
     return createDate;
   }
 
-  public void setCreateDate(Date createDate) {
+  public void setCreateDate(
+                            Date createDate) {
     this.createDate = createDate;
   }
 
   @Override
-  public int getLogonCount() {
+  public long getLogonCount() {
     return logonCount;
   }
 
-  public void setLogonCount(int logonCount) {
+  public void setLogonCount(
+                            long logonCount) {
     this.logonCount = logonCount;
   }
 
   @Override
-  public int getLogonMinutes() {
+  public long getLogonMinutes() {
     return logonMinutes;
   }
 
-  public void setLogonMinutes(int logonMinutes) {
+  public void setLogonMinutes(
+                              long logonMinutes) {
     this.logonMinutes = logonMinutes;
   }
 }
